requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t server/lib/' to install these dependencies
# in `server/lib/` subdirectory.
#
# Note: The `server/lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10.1
requests==2.3
markdown==2.6.4
python-dateutil==2.3
Babel==1.3
# GoogleAppEngineCloudStorageClient==1.9.15.0
# Flask-Security==1.7.4
